fix) free 수정, spte 수정

+++ project3/pintos/src/vm/page.h
@@ -3,7 +3,7 @@
 
 
 struct spage_entry{
-    uint8_t *upage;
+    uint32_t *upage;
 
     struct file *file;
     off_t ofs;
@@ -16,6 +16,6 @@ struct spage_entry{
 
 struct list spage_table;
 
-void spte_create(struct file *file, off_t ofs, uint8_t *upage, uint32_t read_bytes, uint32_t zero_bytes, bool writable);
+void spte_create(struct file *file, off_t ofs, uint32_t *upage, uint32_t read_bytes, uint32_t zero_bytes, bool writable);
 void spt_free(struct list *);
 void update_pte(uint32_t* pd, uint32_t* upage);
